backend/django_service/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
web3
requests
